requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
binance-connector==3.5.1
bitarray==2.8.2
cachetools==5.3.2
ccxt==4.1.65
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect==1.2.0
cytoolz==0.12.2
DataRecorder==3.4.11
DBUtils==3.0.3
decorator==5.1.1
distlib==0.3.7
DownloadKit==2.0.0
DrissionPage==4.0.3.3
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
fake-useragent==1.3.0
Faker==22.2.0
filelock==3.13.1
FlowViewer==0.3.0
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
inflect==7.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
loguru==0.7.0
lru-dict==1.2.0
lxml==4.9.3
mnemonic==0.20
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
parsimonious==0.9.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
protobuf==4.25.0
psutil==5.9.7
py==1.11.0
py-solc-x==2.0.2
pycares==4.4.0
pychrome==0.2.4
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pyproject-api==1.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.0.0
pywatchman==1.4.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
retry==0.9.2
rlp==3.0.0
rpds-py==0.10.6
selenium==4.15.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.50
tldextract==5.1.0
toolz==0.12.0
tox==4.11.3
trio==0.23.0
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.7
web3==6.15.1
webdriver-manager==4.0.0
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
